+++ client-web/src/main.tsx
@@ -14,6 +14,7 @@ import { BlockedUsersRoute } from "./routes/blocked.tsx";
 import { ListsRoute } from "./routes/lists.tsx";
 import { EventRoute } from "./routes/event.tsx";
 import { TagRoute } from "./routes/tag.tsx";
+import { SearchRoute } from "./routes/search.tsx";
 import "./index.css";
 import theme from "./theme.ts";
 
@@ -42,6 +43,7 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
             <Route path="/p/:npub" element={<ProfileRoute />} />
             <Route path="/e/:note" element={<EventRoute />} />
             <Route path="/t/:tag" element={<TagRoute />} />
+            <Route path="/search/:query" element={<SearchRoute />} />
           </Route>
         </Routes>
       </HashRouter>